+++ package.json
@@ -10,16 +10,16 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "json-formatter": "file:",
     "@tailwindcss/vite": "^4.0.9",
     "install": "^0.13.0",
+    "json-formatter": "file:",
     "npm": "^11.1.0",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
-    "react-drag-drop-files": "^2.4.0",
-    "typescript": "^5.8.2",
     "react-icons": "^5.5.0",
-    "tailwindcss": "^4.0.9"
+    "react-router-dom": "^7.2.0",
+    "tailwindcss": "^4.0.9",
+    "typescript": "^5.8.2"
   },
   "devDependencies": {
     "@eslint/js": "^9.21.0",
